base_components/content_embedding/requirements.txt
python-dotenv
pika>=1.3.0
requests